feat(dto): 更新服务ID参数类型为Long，提升数据一致性
- 将ServiceDocDTO和ServiceDocVO中的serviceId参数类型从Integer更改为Long，以确保与数据模型的一致性。
- 更新相关方法的参数类型，确保系统在处理服务ID时的稳定性与兼容性。

这些改动增强了数据模型的健壮性与一致性，确保了整体设计的一致性。

+++ datasophon-common/src/main/java/com/datasophon/common/dto/ServiceDocDTO.java
@@ -32,7 +32,7 @@ import java.io.Serializable;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record ServiceDocDTO(
         Long clusterId,
-        Integer serviceId,
+        Long serviceId,
         String serviceName,
         String docType,
         String docContent,
@@ -46,14 +46,14 @@ public record ServiceDocDTO(
     /**
      * 创建基础文档DTO
      */
-    public static ServiceDocDTO of(Long clusterId, Integer serviceId, String serviceName, String docType) {
+    public static ServiceDocDTO of(Long clusterId, Long serviceId, String serviceName, String docType) {
         return new ServiceDocDTO(clusterId, serviceId, serviceName, docType, null, null, null, "UTF-8");
     }
 
     /**
      * 创建包含内容的文档DTO
      */
-    public static ServiceDocDTO withContent(Long clusterId, Integer serviceId, String serviceName,
+    public static ServiceDocDTO withContent(Long clusterId, Long serviceId, String serviceName,
             String docType, String docContent, String docPath) {
         Long contentLength = docContent != null ? (long) docContent.length() : 0L;
         return new ServiceDocDTO(clusterId, serviceId, serviceName, docType, docContent, docPath, contentLength,
